requirements.txt
openai-whisper
requests
ffmpeg-python